Feat: Adding support for urls in database config (#307)
* Feat: Adding support for urls in database config

This includes both a url key in the database config and
a DATABASE_URL environment variable.

* Updated per recommendations

+++ src/packages/cli/commands/dbdrop.js
@@ -1,7 +1,8 @@
-import { CWD, NODE_ENV } from '../../../constants';
+import { CWD, NODE_ENV, DATABASE_URL } from '../../../constants';
 import { connect } from '../../database';
 import { rmrf } from '../../fs';
 import loader from '../../loader';
+import { CONNECTION_STRING_MESSAGE } from '../constants';
 
 /**
  * @private
@@ -12,6 +13,7 @@ export async function dbdrop() {
       [NODE_ENV]: {
         driver,
         database,
+        url,
         ...config
       }
     }
@@ -20,6 +22,9 @@ export async function dbdrop() {
   if (driver === 'sqlite3') {
     await rmrf(`${CWD}/db/${database}_${NODE_ENV}.sqlite`);
   } else {
+    if (DATABASE_URL || url) {
+      return console.log(CONNECTION_STRING_MESSAGE);
+    }
     const { schema } = connect(CWD, { ...config, driver });
     const query = schema.raw(`DROP DATABASE IF EXISTS ${database}`);
